+++ RELEASE_NOTES.md
@@ -4,7 +4,15 @@
 
 ### New features
 
+- New animations
+- New colors for the policies Cards
+- Filter by resource in the policies table
+
 ### Bug fixes
+
+- Fix inside policies form the default path value
+- Fix inside policies form a missing translation
+- Changing translations inside the policies table
 
 ### Documentation